+++ CHANGES.md
@@ -6,6 +6,7 @@
   points to `Nodes::Attributes` data structures (a replacement for `Activity::NodeAttributes`).
   This decision reduces logic and improves performance: it turned out that most of the time an introspect
   lookup queries for a task, not ID.
+* Re-add `:nodes` to the container activity as this gives a consistent way for treating all `Activity`s.
 * Remove `Activity::NodeAttributes`.
 * Move `Introspect::Graph` to `trailblazer-developer`. It's a data structure very specific
   to rendering, which is not a part of pure runtime behavior. `Activity::Introspect.Graph()` is now deprecated.